yaml deployment files for fold detection and fold dilation

+++ inference/fold_postprocess_batch.py
@@ -12,7 +12,7 @@ from args import get_argparser, parse_args, get_aligner, get_bbox, get_provenanc
 from os.path import join
 from cloudmanager import CloudManager
 from time import time
-from tasks import run 
+from tasks import run
 
 def print_run(diff, n_tasks):
   if n_tasks > 0:
@@ -44,7 +44,7 @@ if __name__ == '__main__':
     help='Value of of mask that indicates DO NOT mask')
   parser.add_argument('--dst_path', type=str)
   parser.add_argument('--mip', type=int)
-  parser.add_argument('--thr_binarize', type=int, default=0,
+  parser.add_argument('--thr_binarize', type=float, default=0,
     help='Threshold for binary mask')
   parser.add_argument('--w_connect', type=int, default=0,
     help='Width to dilate to connect adjacent components')
@@ -59,8 +59,8 @@ if __name__ == '__main__':
   parser.add_argument('--bbox_mip', type=int, default=0,
     help='MIP level at which bbox_start & bbox_stop are specified')
   parser.add_argument('--max_mip', type=int, default=9)
-  parser.add_argument('--max_displacement', 
-    help='the size of the largest displacement expected; should be 2^high_mip', 
+  parser.add_argument('--max_displacement',
+    help='the size of the largest displacement expected; should be 2^high_mip',
     type=int, default=2048)
   parser.add_argument('--block_size', type=int, default=10)
   args = parse_args(parser)
@@ -70,7 +70,7 @@ if __name__ == '__main__':
   # a.device = torch.device('cpu')
   bbox = get_bbox(args)
   provenance = get_provenance(args)
-  
+
   # Simplify var names
   mip = args.mip
   max_mip = args.max_mip
@@ -115,7 +115,7 @@ if __name__ == '__main__':
   ptask = []
   for i in range_list:
       ptask.append(TaskIterator(i))
- 
+
   if a.distributed:
       with ProcessPoolExecutor(max_workers=a.threads) as executor:
           executor.map(remote_upload, ptask)
